Added: Maximum between three float

+++ src/main/java/com/bridgelabz/Maximum.java
@@ -2,8 +2,8 @@ package com.bridgelabz;
 
 
 public class Maximum {
-    public static Integer getmaximum(Integer a, Integer b, Integer c) {
-        Integer max = a;
+    public static Float getmaximum(Float a, Float b, Float c) {
+        Float max = a;
         if (b.compareTo(max) > 0 && b.compareTo(c) > 0)
             max = b;
         else if (c.compareTo(max) > 0)
